# ragdoll/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "A"
  version: "1.0"
  title: "Configuration traceability"
...
paths:
...
  /confs/queryRealConfs:
    post:
      tags:
        - "confs"
      summary: "query the real configuration value in the current hostId node"
      description: "query the real configuration value in the current hostId node"
      operationId: "query_real_confs"
      parameters:
        - in: "body"
          name: "body"
          required: false
          schema:
            $ref: "#/definitions/ConfHost"
      responses:
        "200":
          description: "query total modules successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealConfInfo"
      x-swagger-router-controller: "ragdoll.controllers.confs_controller"
  /confs/querySupportedConfs:
    post:
      tags:
        - "confs"
      summary: "query supported configuration list"
      description: "query supported configuration list"
      operationId: "query_supported_confs"
      parameters:
        - in: "body"
          name: "body"
          description: "domain info"
          required: false
          schema:
            $ref: "#/definitions/DomainName"
      responses:
        "200":
          description: "query supported configuration list successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SupportedConf"
      x-swagger-router-controller: "ragdoll.controllers.confs_controller"
  /confs/queryExpectedConfs:
    post:
      tags:
        - "confs"
      summary: "query expected configuration value in the current hostId node"
      description: "queryExpectedConfs"
      operationId: "query_excepted_confs"
      parameters:
        - in: "body"
          name: "body"
          required: false
          schema:
            $ref: "#/definitions/DomainNames"
      responses:
        "200":
          description: "query expected configuration value successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ExceptedConfInfo"
        "404":
          description: "path does not exist"
      x-swagger-router-controller: "ragdoll.controllers.confs_controller"
  /confs/syncConf:
    put:
      tags:
        - "confs"
      summary: "synchronize the configuration information of the configuration domain\
        \ to the host"
      description: "synchronize the configuration information of the configuration\
        \ domain to the host"
      operationId: "sync_conf_to_host_from_domain"
      parameters:
        - in: "body"
          name: "body"
          required: false
          schema:
            $ref: "#/definitions/SyncReq"
      responses:
        "200":
          description: "synchronize the configuration items successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/HostSyncResult"
      x-swagger-router-controller: "ragdoll.controllers.confs_controller"
  /confs/batch/syncConf:
    put:
      tags:
        - "confs"
      summary: "batch synchronize the configuration information of the configuration domain\
        \ to the host"
      description: "batch synchronize the configuration information of the configuration\
        \ domain to the host"
      operationId: "batch_sync_conf_to_host_from_domain"
      parameters:
        - in: "body"
          name: "body"
          required: false
          schema:
            $ref: "#/definitions/BatchSyncReq"
      responses:
        "200":
          description: "synchronize the configuration items successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/HostSyncResult"
      x-swagger-router-controller: "ragdoll.controllers.confs_controller"
  /confs/domain/diff:
    post:
      tags:
        - "confs"
      summary: "compare domain conf different"
      description: "compare domain conf different"
      operationId: "compare_conf_diff"
      parameters:
        - in: "body"
          name: "body"
          required: false
          schema:
            $ref: "#/definitions/CompareConfDiff"
      responses:
        "200":
          description: "compare domain conf different successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/DomainConfigSyncResult"
      x-swagger-router-controller: "ragdoll.controllers.confs_controller"
  /confs/getDomainStatus:
    post:
      tags:
        - "confs"
      summary: "get the status of the domain"
      description: "get the status of whether the domain has been synchronized"
      operationId: "get_the_sync_status_of_domain"
      parameters:
        - in: "body"
          name: "body"
          required: false
          schema:
            $ref: "#/definitions/DomainIp"
      responses:
        "200":
          description: "get the status of the domain successfully"
          schema:
            $ref: "#/definitions/SyncStatus"
      x-swagger-router-controller: "ragdoll.controllers.confs_controller"
...
definitions:
...
  DomainName:
    type: "object"
    properties:
      domainName:
        type: "string"
        description: "domain name"
  DomainNames:
    type: "object"
    properties:
      domainNames:
        type: "array"
        items:
          $ref: "#/definitions/DomainName"
  DomainIp:
    type: "object"
    properties:
      domainName:
        type: "string"
        description: "domain name"
      ip:
        type: "string"
        description: "ip"
...
  ConfHost:
    type: "object"
    properties:
      domainName:
        type: "string"
        description: "domain name"
      hostIds:
        type: "array"
        items:
          type: "object"
          properties:
            hostId:
              type: "integer"
              description: "the id of host"
  RealConfInfo:
    type: "object"
    properties:
      domainName:
        type: "string"
      hostID:
        type: "integer"
      confBaseInfos:
        type: "array"
        items:
          $ref: "#/definitions/RealconfBaseInfo"
    example:
      domainName: "domainName"
      hostID: "hostID"
      confBaseInfos:
        - rpmName: "rpmName"
          path: "path"
          filePath: "filePath"
          spacer: "spacer"
          rpmVersion: "rpmVersion"
          rpmRelease: "rpmRelease"
          fileOwner: "fileOwner"
          confType: "key-value"
          confContents: "confContents"
          fileAttr: "fileAttr"
  RealconfBaseInfo:
    type: "object"
    properties:
      path:
        type: "string"
        description: "The currently configured xpath path"
      filePath:
        type: "string"
        description: "The real file path of the configuration item in OS"
      rpmName:
        type: "string"
      rpmVersion:
        type: "string"
      rpmRelease:
        type: "string"
      fileAttr:
        type: "string"
        description: "the attributes for the real file of the configuration item in\
          \ OS"
      fileOwner:
        type: "string"
        description: "the owner information for the real file of the configuration\
          \ item in OS"
      confType:
        type: "string"
        description: "the configuration type"
        enum:
          - "key-value"
          - "toml"
          - "json"
          - "xml"
          - "text"
      spacer:
        type: "string"
        description: "spacers of configuration items in the configuration file"
      confContents:
        type: "string"
        description: "the specific content of the configuration item"
    example:
      rpmName: "rpmName"
      path: "path"
      filePath: "filePath"
      spacer: "spacer"
      rpmVersion: "rpmVersion"
      rpmRelease: "rpmRelease"
      fileOwner: "fileOwner"
      confType: "key-value"
      confContents: "confContents"
      fileAttr: "fileAttr"
  ExceptedConfInfo:
    type: "object"
    properties:
      domainName:
        type: "string"
      confBaseInfos:
        type: "array"
        items:
          $ref: "#/definitions/ConfBaseInfo"
    example:
      domainName: "domainName"
      confBaseInfos:
        - expectedContents: "expectedContents"
          filePath: "filePath"
          changeLog:
            - date: "2000-01-23T04:56:07.000+00:00"
              preValue: "preValue"
              changeReason: "changeReason"
              author: "author"
              postValue: "postValue"
              changeId: "changeId"
            - date: "2000-01-23T04:56:07.000+00:00"
              preValue: "preValue"
              changeReason: "changeReason"
              author: "author"
              postValue: "postValue"
              changeId: "changeId"
        - expectedContents: "expectedContents"
          filePath: "filePath"
          changeLog:
            - date: "2000-01-23T04:56:07.000+00:00"
              preValue: "preValue"
              changeReason: "changeReason"
              author: "author"
              postValue: "postValue"
              changeId: "changeId"
            - date: "2000-01-23T04:56:07.000+00:00"
              preValue: "preValue"
              changeReason: "changeReason"
              author: "author"
              postValue: "postValue"
              changeId: "changeId"
  ConfBaseInfo:
    type: "object"
    properties:
      filePath:
        type: "string"
        description: "the path of a configuration file"
      expectedContents:
        type: "string"
        description: "expected configuration value of configuration item"
      changeLog:
        type: "array"
        items:
          $ref: "#/definitions/GitLogMessage"
    example:
      expectedContents: "expectedContents"
      filePath: "filePath"
      changeLog:
        - date: "2000-01-23T04:56:07.000+00:00"
          preValue: "preValue"
          changeReason: "changeReason"
          author: "author"
          postValue: "postValue"
          changeId: "changeId"
        - date: "2000-01-23T04:56:07.000+00:00"
          preValue: "preValue"
          changeReason: "changeReason"
          author: "author"
          postValue: "postValue"
          changeId: "changeId"
  GitLogMessage:
    type: "object"
    properties:
      date:
        type: "string"
        format: "date-time"
      author:
        type: "string"
      changeId:
        type: "string"
      changeReason:
        type: "string"
      preValue:
        type: "string"
      postValue:
        type: "string"
    example:
      date: "2000-01-23T04:56:07.000+00:00"
      preValue: "preValue"
      changeReason: "changeReason"
      author: "author"
      postValue: "postValue"
      changeId: "changeId"
...
  SyncStatus:
    type: "object"
    properties:
      domainName:
        type: "string"
        description: "domain name"
      hostStatus:
        type: "array"
        items:
          $ref: "#/definitions/HostSyncStatus"
    example:
      hostStatus:
        - hostId: "hostId"
          syncStatus:
            - isSynced: "NOT FOUND"
              filePath: "filePath"
            - isSynced: "NOT FOUND"
              filePath: "filePath"
        - hostId: "hostId"
          syncStatus:
            - isSynced: "NOT FOUND"
              filePath: "filePath"
            - isSynced: "NOT FOUND"
              filePath: "filePath"
      domainName: "domainName"
  HostSyncStatus:
    type: "object"
    properties:
      hostId:
        type: "integer"
        description: "the id of host"
      syncStatus:
        type: "array"
        items:
          $ref: "#/definitions/ConfIsSynced"
    example:
      hostId: "hostId"
      syncStatus:
        - isSynced: "NOT FOUND"
          filePath: "filePath"
        - isSynced: "NOT FOUND"
          filePath: "filePath"
  ConfIsSynced:
    type: "object"
    properties:
      filePath:
        type: "string"
      isSynced:
        type: "string"
        enum:
          - "NOT FOUND"
          - "NOT SYNCHRONIZE"
          - "SYNCHRONIZED"
      singleConf:
        type: "object"
        items:
          $ref: "#/definitions/SingleConf"
    example:
      isSynced: "NOT FOUND"
      filePath: "filePath"
      singleConf: "singleConf"
  HostSyncResult:
    type: "object"
    properties:
      hostId:
        type: "integer"
        description: "the id of host"
      syncStatus:
        type: "string"
        description: "the result of host"
        enum:
          - "SUCCESS"
          - "FAILED"
    example:
      hostId: "hostId"
      syncStatus: "SUCCESS"
  DomainConfigSyncResult:
    type: "object"
    properties:
      hostId:
        type: "integer"
        description: "the id of host"
      domainName:
        type: "string"
        description: "the domainName of host"
      syncStatus:
        type: "integer"
        description: "the syncStatus of host"
...
  SupportedConf:
    type: "string"
  SyncHostConfs:
    type: "object"
    properties:
      hostId:
        type: "integer"
        description: "the id of host"
      syncConfigs:
        type: "array"
        items:
          type: "string"
  CompareConfDiff:
    type: "object"
    properties:
      expectedConfsResp:
        type: "array"
        items:
          $ref: "#/definitions/DomainConfBaseInfos"
      domainResult:
        type: object
        description: "domain real config"
  DomainConfBaseInfos:
    type: "object"
    properties:
      confBaseInfos:
        type: array
        items:
          $ref: "#/definitions/ConfBase"
      domainName:
        type: "string"
        description: "domain name"
  ConfBase:
    type: "object"
    properties:
      filePath:
        type: "string"
        description: "file path"
      expectedContents:
        type: "string"
        description: "expected contents"
  SyncReq:
    type: "object"
    properties:
      domainName:
        type: "string"
        description: "domain name"
      syncList:
        type: "array"
        items:
          $ref: "#/definitions/SyncHostConfs"
  BatchSyncReq:
    type: "object"
    properties:
      domainName:
        type: "string"
        description: "domain name"
      hostIds:
        type: "array"
        items:
          type: "integer"
  SingleConf:
    type: object
    properties:
      singleFilePath:
        type: "string"
        description: "single file path"
      singleIsSynced:
        type: "string"
        enum:
          - "NOT FOUND"
          - "NOT SYNCHRONIZE"
          - "SYNCHRONIZED"